requirements.txt
google-generativeai
numpy
pytest
ipykernel
python-dotenv
pydantic
pypdf
pandas
causallearn
pyvis
